requirements.txt
customtkinter
gitpython
openai
requests
python-dotenv
pytest
tiktoken
